==> scripts/2b) map obs alpha diversity.R <==


# MAP ALPHADIVERSITY FOR OBSERVED ASSEMBLAGES


source('scripts/1) import and prepare data.R')
library(SpatialPack) # modified.ttest


# data
obs_alpha_div <- read.csv("results/obs_alpha_div.txt", sep="")
world_grid



# dataframe to map #### 

# convert n_ann into proportion
obs_alpha_div$r_habit <- c(obs_alpha_div$richness-obs_alpha_div$n_ann)/obs_alpha_div$n_ann

# template to fill
alpha_div_rasters <- rep(world_grid, 13)
names(alpha_div_rasters) <- c("richness", "cwm_height", "cwm_blade", "cwm_nutlet", "r_habit",
                              "Frich", "Fmpd", "Prich","Pmpd","perc_Frich", "perc_Fmpd", "perc_Prich", "perc_Pmpd")
                  
# fill values
alpha_div_rasters$richness <- terra::rast(obs_alpha_div[,c('x','y','richness')], crs='+proj=eqearth') %>%
    terra::extend(world_grid)

alpha_div_rasters$cwm_height <- terra::rast(obs_alpha_div[,c('x','y','cwm_height')], crs='+proj=eqearth') %>%
  terra::extend(world_grid)
alpha_div_rasters$cwm_blade <- terra::rast(obs_alpha_div[,c('x','y','cwm_blade')], crs='+proj=eqearth') %>%
  terra::extend(world_grid)
alpha_div_rasters$cwm_nutlet <- terra::rast(obs_alpha_div[,c('x','y','cwm_nutlet')], crs='+proj=eqearth') %>%
  terra::extend(world_grid)
alpha_div_rasters$r_habit <- terra::rast(obs_alpha_div[,c('x','y','r_habit')], crs='+proj=eqearth') %>%
  terra::extend(world_grid)

alpha_div_rasters$Frich <- terra::rast(obs_alpha_div[,c('x','y','Frich')], crs='+proj=eqearth') %>%
  terra::extend(world_grid)
alpha_div_rasters$Fmpd <- terra::rast(obs_alpha_div[,c('x','y','Fmpd')], crs='+proj=eqearth') %>%
  terra::extend(world_grid)
alpha_div_rasters$Prich <- terra::rast(obs_alpha_div[,c('x','y','Prich')], crs='+proj=eqearth') %>%
  terra::extend(world_grid)
alpha_div_rasters$Pmpd <- terra::rast(obs_alpha_div[,c('x','y','Pmpd')], crs='+proj=eqearth') %>%
  terra::extend(world_grid)

alpha_div_rasters$perc_Frich <- terra::rast(obs_alpha_div[,c('x','y','perc_Frich')], crs='+proj=eqearth') %>%
  terra::extend(world_grid)
alpha_div_rasters$perc_Fmpd <- terra::rast(obs_alpha_div[,c('x','y','perc_Fmpd')], crs='+proj=eqearth') %>%
  terra::extend(world_grid)
alpha_div_rasters$perc_Prich <- terra::rast(obs_alpha_div[,c('x','y','perc_Prich')], crs='+proj=eqearth') %>%
  terra::extend(world_grid)
alpha_div_rasters$perc_Pmpd <- terra::rast(obs_alpha_div[,c('x','y','perc_Pmpd')], crs='+proj=eqearth') %>%
  terra::extend(world_grid)

# save rasters
writeRaster(alpha_div_rasters, 'results/maps/obs_alpha_div_rasters.tiff', overwrite=TRUE)



# prepare maps for representation ####

for (s in 1:dim(alpha_div_rasters)[3]) {
  
  pdf(file = paste("results/maps/alpha_div_observed_50/",names(alpha_div_rasters)[s],".pdf",sep=""),
      width = 9.30, # The width of the plot in inches
      height = 5.74) # The height of the plot in inches
  alpha_div_rasters[[s]] %>%
    plot(main=names(alpha_div_rasters)[s], col=rev(grDevices::heat.colors(20)[1:15]))
  lines(world_lines)
  
  dev.off()
  
}


# SES
par(mfrow=c(2,2))

temp <- alpha_div_rasters$perc_Frich
temp[temp>=0.95] <- 1; temp[temp<=0.05] <- 2; temp[temp<0.95 & temp>0.05] <- 3
plot(temp, col=c('indianred','blue','snow3'), legend=NULL, main='Significant Frich')
lines(world_lines, alpha=0.5)

temp <- alpha_div_rasters$perc_Fmpd
temp[temp>=0.95] <- 1; temp[temp<=0.05] <- 2; temp[temp<0.95 & temp>0.05] <- 3
plot(temp, col=c('indianred','blue','snow3'), legend=NULL, main='Significant Fmpd')
lines(world_lines, alpha=0.5)

temp <- alpha_div_rasters$perc_Prich
temp[temp>=0.95] <- 1; temp[temp<=0.05] <- 2; temp[temp<0.95 & temp>0.05] <- 3
plot(temp, col=c('indianred','blue','snow3'), legend=NULL, main='Significant Prich')
lines(world_lines, alpha=0.5)

temp <- alpha_div_rasters$perc_Pmpd
temp[temp>=0.95] <- 1; temp[temp<=0.05] <- 2; temp[temp<0.95 & temp>0.05] <- 3
plot(temp, col=c('indianred','blue','snow3'), legend=NULL, main='Significant Pmpd')
lines(world_lines, alpha=0.5)
